Make asset lookups relative to project directory
When run in development, the current working directory might not be
the root of this repo.  If this is the case, finding assets relative
to the root of this repo will allow the server to serve the static
assets correctly.

This naive patch attempts to make all asset lookups relative to the
root of this repo so that a controlling process can monitor all
development daemons from another workin directory.