MOCK is not set, it should not attempt to MOCK
Setting MOCK as an environment variable (to anything) should invoke
detection of MOCK logic.  Not setting MOCK should not use MOCK logic.

As for the value of MOCK, I would say setting it to anything other
than empty string should use MOCK data, but I will leave the existing
logic to only use MOCK when MOCK is set to "true".