Fix underscore dependency for tests and browserify
Tideline is a client-side lib that expects a global `window` object (to
grap its dependencies, like underscore).

Tests (for now?) are run in Node (vs. the browser), which means no
global `window` object.

We manually add `global.window` with necessary dependencies before
running the tests.

Note that we can't just use `require('underscore')`, as this will bundle
all of underscore into `tideline.js` (or just produce an error as
underscore isn't installed in `node_modules` when installing Tideline
with Bower).