Fix most recent function on chart
Problem: The `datetimeLocation` prop on ChartDaily, would determine the
call to `chart.locate()`. If you switched to weekly view, that would set
the datetime location to some value, when switching back to daily view,
you would be at that location (which is what you want), but then
clicking "most recent" would not do anything as the datetime location
was set to some value.

Fix: Separate chart initialization (data, and locating), into its own
function with parameters, and explicity call it with no datetime
location when user clicks "most recent".