Browser-side data conversion code for blip
This provides some code for blip to convert data as it comes from the tidepool API into what it wants for visualization.  I cannot figure out how to get the tests to actually run nor can I figure out how to properly package the code yet, so this is an incomplete commit that is being pushed up for feedback on what I'm doing wrong.

There are also some fixes to pass jshint