Handle boluses with an unknown split
This change adjust the bolus rendering code to be able to handle boluses with an unknown immediate and extended split.  To get diasend data working properly though, it requires the diasend data to be adjusted some, which is currently being done in blip.  Without the blip changes, the rendering will remain the same, with the blip changes, diasend square basals will have a duration line on them.