Change stats annotations for cgm vs smbg
The puddle objects for Range and Average now store a base
annotation key string, ie 'stats-how-calculated-range', and when the
annotation is actually built, '-cgb' or '-smbg' is appended for
looking up the corresponding annotation content string.

One key change I made was no longer using the default annotationOpts
in conjunction with _.defaults. The reason for this is that it can
introduce subtle bugs, because apparently the default objects are
then shared by any objects returned by _.defaults. The quick fix
was simply to build separate annotationOpts objects for each puddle
and set up the defaults individually, which isn't too bad.