Fix bug with building app.config
Basically Webpack CLI tool --entry option does not respect that the configured entry option could be an array. In this instance the option supplied from the command line is completely ignored and you get out the default entry point(s). Luckily as we only have one entry point when in production (when we build config) we can change from specifying it as an array with one value, to a string!