Account for patient being other person
When displaying a list of patients whose data you are able to view we need to ensure the list is sorted by the patient's name rather than the name associated with the profile that manages the patient. This commit updates the sorting rules and adds a new test condition to ensure that this property is used when the patient isOtherPerson flag is set to true

Remove console.log

Missing comma in test set-up!

Missing comma in test set-up!