PT-17 Saving user profile may trigger a race condition with seagull. (#74)
When saving clinician profile with phone number Blip tries to update the profile and the preferences at the same time. Both are calls to Seagull and this happens simultaneously which causes a race condition.
In this situation MongoDB send back an error: The userId is unique!

Solution was to change the parallel calls to series.